Pair internal/handlers logger test with a split source file
